requirements.txt
openai-whisper
fastapi
uvicorn
python-multipart